# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Leumas Defensive Cyber Agent API
  version: 0.1.0
  description: Local-first defensive cyber triage, case management, safety, and operator readiness API.
servers:
  - url: http://127.0.0.1:3001
    description: Local development API
...
paths:
...
  /cases/{id}:
    get:
      tags: [cases]
      operationId: getCase
      summary: Fetch a stored investigation case.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stored case record.
        "404":
          description: Case was not found.
...
  /connectors/{connectorId}/query:
    post:
      tags: [tools]
      operationId: querySecurityConnector
      summary: Run a read-only connector query without importing records to a case.
      parameters:
        - name: connectorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Read-only connector query result.
        "403":
          description: Connector policy denied the query.
  /cases/{id}/connectors/{connectorId}/import:
    post:
      tags: [tools]
      operationId: importConnectorEvidenceToCase
      summary: Run a read-only connector query and store evidence references on a case.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: connectorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Imported connector evidence references and updated case.
        "403":
          description: Connector policy denied the import.
...